requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
openai>=1.0.0
aiohttp>=3.9.0
asyncio>=3.4.3
sqlalchemy>=2.0.0
alembic>=1.13.0
pandas>=2.0.0
numpy>=1.24.0
tiktoken>=0.5.0
rich>=13.0.0
click>=8.1.0
pyyaml>=6.0
aiofiles>=23.0.0
letta>=0.3.0
chromadb>=0.4.0
langchain>=0.1.0 